register_login_ms/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
Flask-RESTful==0.3.10
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.7.3
pytz==2024.1
requests==2.32.3
six==1.16.0
urllib3==2.2.2
Werkzeug==3.0.3